main file
Was working last time I checked, though is the cert changed and stopped the other one from working, this may not work now too. 

The only issue with this at the time, was lack of autofit columns and that the website itself is unreliable. often parking is just a - and it is written in the description that there is or is not. I could scrape for that, but different agents also word it differently so would be a bit much just for this practise project (i felt).